Fixed INotifyCollectionChanged support.
Changed DataBinding to work with DataTemplates and no longer as partial classes to the elements.

Totally rewritten the MultilineElement to grow the Cell as you type (Lines = 0) or
restricted to a set number of lines. (Still some redraw issues)

DateTimeElements are no longer working in this release. I am pushing this version up so that Travis can get his code working. Don't get this one if you need Date or Time Elements