Formatting changes, Added IActivation for Views to hook into Activation and Deactivation, Added ElementPropertyName and InstancePropertyName to most Attributes
Attributes that create Elements or UIViews can now take a property name and will fill that property with the element and / or Instance of the view that was created. You can use this is you need to get at the actual views and elements that make up your view.

Also added a DataView property to Root that holds an instance of the View that was used to make up the TableView.