Fixed NotifyCollectionChanged handler in DataContextBinder when updating items that are not currently displayed.
Sometimes when setting a collection based from another dependent collection like Master - detail changing the master will cause the detail to be changed. In this case DataContextBinder now checks to see if the controller is has a reference to is the same controller that is the current controller. If not it won't try to update the tableview.